+++ crates/fuel-core/src/graphql_api/database.rs
@@ -35,6 +35,7 @@ use fuel_core_types::{
         block::CompressedBlock,
         consensus::Consensus,
         primitives::{
+            BlockHeightQuery,
             BlockId,
             DaBlockHeight,
         },
@@ -156,42 +157,46 @@ impl DatabaseBlocks for ReadView {
 
     fn blocks(
         &self,
-        height: Option<BlockHeight>,
+        height: BlockHeightQuery,
         direction: IterDirection,
     ) -> BoxedIter<'_, StorageResult<CompressedBlock>> {
         // Chain together blocks from the off-chain db and the on-chain db
         // The blocks in off-chain db, if any, are from time before regenesis
 
-        if let Some(height) = height {
-            match (height >= self.genesis_height, direction) {
-                (true, IterDirection::Forward) => {
-                    self.on_chain.blocks(Some(height), direction)
-                }
+        if let BlockHeightQuery::Specific(inner) = height {
+            match (inner >= self.genesis_height, direction) {
+                (true, IterDirection::Forward) => self.on_chain.blocks(height, direction),
                 (true, IterDirection::Reverse) => self
                     .on_chain
-                    .blocks(Some(height), direction)
-                    .chain(self.off_chain.old_blocks(None, direction))
+                    .blocks(height, direction)
+                    .chain(
+                        self.off_chain
+                            .old_blocks(BlockHeightQuery::Genesis, direction),
+                    )
                     .into_boxed(),
                 (false, IterDirection::Forward) => self
                     .off_chain
-                    .old_blocks(Some(height), direction)
-                    .chain(self.on_chain.blocks(None, direction))
+                    .old_blocks(height, direction)
+                    .chain(self.on_chain.blocks(BlockHeightQuery::Genesis, direction))
                     .into_boxed(),
                 (false, IterDirection::Reverse) => {
-                    self.off_chain.old_blocks(Some(height), direction)
+                    self.off_chain.old_blocks(height, direction)
                 }
             }
         } else {
             match direction {
                 IterDirection::Forward => self
                     .off_chain
-                    .old_blocks(None, direction)
-                    .chain(self.on_chain.blocks(None, direction))
+                    .old_blocks(BlockHeightQuery::Genesis, direction)
+                    .chain(self.on_chain.blocks(BlockHeightQuery::Genesis, direction))
                     .into_boxed(),
                 IterDirection::Reverse => self
                     .on_chain
-                    .blocks(None, direction)
-                    .chain(self.off_chain.old_blocks(None, direction))
+                    .blocks(BlockHeightQuery::Genesis, direction)
+                    .chain(
+                        self.off_chain
+                            .old_blocks(BlockHeightQuery::Genesis, direction),
+                    )
                     .into_boxed(),
             }
         }
@@ -329,7 +334,7 @@ impl OffChainDatabase for ReadView {
 
     fn old_blocks(
         &self,
-        height: Option<BlockHeight>,
+        height: BlockHeightQuery,
         direction: IterDirection,
     ) -> BoxedIter<'_, StorageResult<CompressedBlock>> {
         self.off_chain.old_blocks(height, direction)