+++ crates/fuel-core/src/service/adapters/graphql_api.rs
@@ -29,6 +29,7 @@ use fuel_core_txpool::{
     types::TxId,
 };
 use fuel_core_types::{
+    blockchain::primitives::BlockHeightQuery,
     entities::relayer::message::MerkleProof,
     fuel_tx::{
         Bytes32,
@@ -201,7 +202,7 @@ impl worker::BlockImporter for GraphQLBlockImporter {
 
     fn block_event_at_height(
         &self,
-        height: Option<BlockHeight>,
+        height: BlockHeightQuery,
     ) -> anyhow::Result<SharedImportResult> {
         self.import_result_provider_adapter.result_at_height(height)
     }