+++ crates/types/src/blockchain/primitives.rs
@@ -3,7 +3,10 @@
 use crate::{
     fuel_crypto,
     fuel_crypto::SecretKey,
-    fuel_types::Bytes32,
+    fuel_types::{
+        BlockHeight,
+        Bytes32,
+    },
 };
 use core::array::TryFromSliceError;
 use derive_more::{
@@ -176,5 +179,32 @@ impl TryFrom<&'_ [u8]> for BlockId {
 
     fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
         Ok(Self::from(TryInto::<[u8; 32]>::try_into(bytes)?))
+    }
+}
+
+/// Represents either the Genesis Block or a block at a specific height
+#[derive(Copy, Clone, Debug, Display, PartialEq, Eq, Hash, Ord, PartialOrd)]
+pub enum BlockHeightQuery {
+    /// Block at a specific height
+    Specific(BlockHeight),
+    /// Genesis block
+    Genesis,
+}
+
+impl From<BlockHeightQuery> for Option<BlockHeight> {
+    fn from(value: BlockHeightQuery) -> Self {
+        match value {
+            BlockHeightQuery::Specific(block_height) => Some(block_height),
+            BlockHeightQuery::Genesis => None,
+        }
+    }
+}
+
+impl From<Option<BlockHeight>> for BlockHeightQuery {
+    fn from(value: Option<BlockHeight>) -> Self {
+        match value {
+            Some(block_height) => BlockHeightQuery::Specific(block_height),
+            None => BlockHeightQuery::Genesis,
+        }
     }
 }